Code restructure: moving compilation driving logic to Driver rather than in main
